Final/requirements.txt
tqdm
python-magic
lief
capstone
keystone